shop_payment_service/requirements.txt
asgiref
Django
djangorestframework
Pillow
pytz
sqlparse
typing_extensions
pytest
pytest-django
requests
django-cors-headers
coreapi
coreschema
stripe

# payments linst
https://scalereal.com/backend/2021/12/20/razorpay-payment-gateway-integration-with-django.html
https://studygyaan.com/django/django-paypal-payment-gateway-integration-tutorial
https://developer.paypal.com/braintree/docs/guides/transactions/python
https://django-payments.readthedocs.io/en/latest/backends.html
https://episyche.com/blog/how-to-integrate-razorpay-payment-gateway-in-django-and-react-for-the-checkout-use-case
 # important
https://medium.com/@hari154542/how-to-integrate-stripe-payment-gateway-in-django-rest-framework-232ee341601